+++ src/main/java/goingto/com/controller/sprint4/LocatableReviewsController.java
@@ -1,9 +1,9 @@
-package goingto.com.controller;
+package goingto.com.controller.sprint4;
 
 import goingto.com.model.geographic.Locatable;
-import goingto.com.resource.converter.TipConverter;
+import goingto.com.resource.converter.ReviewConverter;
 import goingto.com.service.LocatableService;
-import goingto.com.service.TipService;
+import goingto.com.service.ReviewService;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,24 +18,23 @@ import java.util.stream.Collectors;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
-public class LocatableTipsController {
+public class LocatableReviewsController {
 
     @Autowired
-    private TipService tipService;
+    private ReviewService reviewService;
     @Autowired
     private LocatableService locatableService;
     @Autowired
-    private TipConverter mapper;
+    private ReviewConverter mapper;
 
-    @ApiOperation("Return Tips by Locatable id")
-    @GetMapping("/locatables/{locatableId}/tips")
-    public ResponseEntity<?> getAllTipsByLocatableId(@PathVariable Integer locatableId){
+    @ApiOperation("Return Reviews by Locatable id")
+    @GetMapping("/locatables/{locatableId}/reviews")
+    public ResponseEntity<?> getAllReviewsByLocatableId(@PathVariable Integer locatableId){
         Locatable existingLocatable = locatableService.getLocatable(locatableId);
         if(existingLocatable==null)
             return ResponseEntity.notFound().build();
-        var tips = tipService.getAllTipsByLocatableId(locatableId);
-        var result = tips.stream().map(mapper::convertToResource).collect(Collectors.toList());
+        var reviews = reviewService.getAllReviewsByLocatableId(locatableId);
+        var result = reviews.stream().map(mapper::convertToResource).collect(Collectors.toList());
         return ResponseEntity.ok(result);
     }
-
 }
